refactor: Move McStopArrival into c1 package
This prepare for making McStopArrival a seald class.
