+++ raptor/src/test/java/org/opentripplanner/raptor/rangeraptor/multicriteria/arrivals/c1/ArrivalParetoSetComparatorFactoryTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals;
+package org.opentripplanner.raptor.rangeraptor.multicriteria.arrivals.c1;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
